Remove duplicate functors and remove unnecessary include in tuple_impl.h
