+++ include/oneapi/dpl/pstl/tuple_impl.h
@@ -21,8 +21,6 @@
 #include <cassert>
 #include <type_traits>
 
-#include "utils.h"
-
 namespace oneapi
 {
 namespace dpl
